+++ inst/doc/O1_make_research_project.R
@@ -1,0 +1,13 @@
+## ---- include = FALSE---------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width=6, 
+  fig.height=4
+)
+
+options(rmarkdown.html_vignette.check_title = FALSE)
+
+## -----------------------------------------------------------------------------
+sessionInfo()
+