+++ inst/doc/O2_make_research_project_markdown.R
@@ -1,0 +1,11 @@
+## ---- include = FALSE---------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+options(rmarkdown.html_vignette.check_title = FALSE)
+
+## -----------------------------------------------------------------------------
+sessionInfo()
+